feat: 문자 개수와 유효한 문자만 들어왔는지 검사
