+++ srcs/parse_map/Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: jayoon <jayoon@student.42seoul.kr>         +#+  +:+       +#+         #
+#    By: jayoon <jayoon@student.42.fr>              +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/04/27 15:13:41 by kipark            #+#    #+#              #
-#    Updated: 2022/11/05 03:51:52 by jayoon           ###   ########.fr        #
+#    Updated: 2022/11/07 21:12:26 by jayoon           ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -40,7 +40,8 @@ SRCS_PARSE_MAP				:=	init_map_info.c			\
 								get_cnt_map_height.c	\
 								get_cnt_map_width.c		\
 								pass_empty_line.c		\
-								pass_element.c			
+								pass_element.c			\
+								get_map.c
 
 SRCS_ERROR					:=	error.c
