+++ srcs/parse_map/asset/example.cub
@@ -14,5 +14,5 @@ F 220,100,0
 
 EA ./path_to_the_east_texture
 
-1
-1
+1W
+1S
